Migrate to null safety
